+++ cookbooks/elastic-stack/recipes/kibana.rb
@@ -1,4 +1,3 @@
-
 # Cookbook Name:: elastic-stack
 # Recipe:: kibana
 # Author:: Wazuh <info@wazuh.com>
@@ -44,18 +43,15 @@ end
 
 # Install the Wazuh Kibana plugin
 
-bash 'Install the Wazuh app plugin for Kibana' do
-  code <<-EOH
-    cd #{node['kibana']['package_path']}
-    sudo -u kibana bin/kibana-plugin install https://packages.wazuh.com/#{node['wazuh']['major_version']}/ui/kibana/wazuh_kibana-#{node['wazuh']['kibana_plugin_version']}-1.zip
-  EOH
+execute 'Install the Wazuh app plugin for Kibana' do
+  command "sudo -u kibana #{node['kibana']['package_path']}/bin/kibana-plugin install https://packages.wazuh.com/#{node['wazuh']['major_version']}/ui/kibana/wazuh_kibana-#{node['wazuh']['kibana_plugin_version']}-1.zip"
 end
 
-# Set up Kibana configuration file
+# Create Kibana configuration file
 
 template "#{node['kibana']['config_path']}/kibana.yml" do
   source 'kibana.yml.erb'
-  owner 'root'
+  owner 'kibana'
   group 'kibana'
   mode 0755
   variables({
@@ -65,22 +61,29 @@ template "#{node['kibana']['config_path']}/kibana.yml" do
   })
 end
 
-bash 'Configure the credentials to access the Wazuh API' do
-  code <<-EOH
-    cat >> #{node['kibana']['package_path']}/optimize/wazuh/config/wazuh.yml << EOF
-    - #{node['kibana']['wazuh_api_credentials']['id']}:
-       url: #{node['kibana']['wazuh_api_credentials']['url']}
-       port: #{node['kibana']['wazuh_api_credentials']['port']}
-       username: #{node['kibana']['wazuh_api_credentials']['username']}
-       password: #{node['kibana']['wazuh_api_credentials']['password']}
-    EOF
-  EOH
-end
-
 # Enable and start the Kibana service
 
 service "kibana" do
   supports :start => true, :stop => true, :restart => true, :reload => true
+  action [:enable, :start]
+end
+
+# Create Wazuh-Kibana plugin configuration file
+
+template "#{node['kibana']['package_path']}/optimize/wazuh/config/wazuh.yml" do
+  source 'wazuh.yml.erb'
+  owner 'kibana'
+  group 'kibana'
+  mode 0755
+  action :create
+  variables ({
+    api_credentials: node['kibana']['wazuh_api_credentials']
+  })
+end
+
+# Restart Kibana service
+
+service "kibana" do
   action [:restart]
 end
 
@@ -94,11 +97,28 @@ ruby_block 'Wait for elasticsearch' do
   end
 end
 
-bash 'Waiting for elasticsearch curl response...' do
+ruby_block 'Wait for kibana' do
+  block do
+    loop { break if (TCPSocket.open(
+      "#{node['kibana']['yml']['server']['host']}",
+      node['kibana']['yml']['server']['port']) rescue nil); 
+      puts "Waiting kibana...."; sleep 60
+    }
+  end
+end
+
+bash 'Waiting for kibana curl response...' do
   code <<-EOH
-  until (curl -XGET #{node['kibana']['yml']['elasticsearch']['hosts']}); do
-    printf 'Waiting for elasticsearch....'
+  until (curl -XGET http://#{node['kibana']['yml']['server']['host']}:#{node['kibana']['yml']['server']['port']}); do
+    printf 'Waiting for kibana....'
     sleep 5
   done
   EOH
+end
+
+log 'Access Kibana web interface' do
+  message "URL: http://#{node['kibana']['yml']['server']['host']}:#{node['kibana']['yml']['server']['port']}
+  user: admin
+  password: admin"
+  level :info
 end