+++ cookbooks/elastic-stack/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## v0.1.0
-1. Bump to Opendistro for Elasticsearch 7.9.1
-2. Added suse repository
+1. Install Elastic Stack 7.9.3 version
+2. Install Wazuh Kibana plugin 4.0.2_7.9.3 version
 *Note*: in future version, kibana will be in an independent cookbook
 
 * Initial/current release